feat: subsume EditCollectionPage under EditPage
This commit modifies the EditPage component so that it subsumes the
EditCollectionPage component under it. To do so, a helper function
is created which generates the CMS api endpoint to hit on the backend,
depending on whether the page in question is a collection page, resource
page, or a normal page. The three types of pages are distinguished
using two props, isResourcePage, and the newly-introduced isCollectionPage.

To accommodate collection page functionality, the EditPage runs an optional
code block which fetches the other pages in a collection (if a page is
a collection page) to display the left navigation bar. The EditPage chooses
to display a SimplePage or a LeftNavPage depending on whether the page
selected is a collection page, and only if there are more than 0 left nav
pages retrieved.

The EditPage component was updated to ensure that the isResourcePage and
isCollectionPage proptypes are compulsory.

And finally, minor stylistic changes were made to ensure that the
SimpleMDE editor stays at max height, instead of shrinking for some
unknown reason. This was done by setting the height of both the SimpleMDE
component and the CodeMirror class to 100%.