feat: create util function to parse collection page filename
This commit creates a util function, prettifyCollectionPageFileName,
to parse collection page file names. Collection page file names are
different from normal file names because the order of collection pages
within a collection are important (they determine the nav layout), so we
need to remove the additional number markers that mark the start of
a collection page.