fix: replace inline styles on LeftNavPage
Previously, the LeftNavPage template used inline styles to forcibly
shift page elements. This resulted in a collision between the page
content html and the left navbar element, since the left navbar was
placed by manually shifting the element to the right.

This commit fixes that by removing the inline styles and reusing the
page-content-body CSS class instead.