fix: remove links from left navigation bar in EditPage
Previously, elements in the left navigation bar for a collection page
in the EditPage component linked to nonsense, non-existent pages on a new
EditPage instance. This is bad behavior and clearly a bug since we don't
want to allow users to edit on non-existent pages.

This commit fixes that by removing all links from the left navigation bar.
The reason we choose to do this instead of making the left navigation bar
unclickable is because we want to EditPage component to retain some sort
of dynamism. However, this might still be confusing to users (having a
glove cursor, and mouseover highlights on an element which is not actually
a link) so we can consider making this unclickable in the future.