feat: revamp left navigation bar on LeftNav template
Previously, the left navigation bar accordion inside the LeftNav
template did not render third nav elements. This means that we needed
to both modify the html template, as well as to use javascript to
create the dynamic accordion behavior of hiding and displaying
accordion child elements.

This commit makes those changes by introducing functions to toggle DOM
element class names (to toggle the visibility of the accordion elements)
and to generate the HTML itself.

However, in doing so, we removed the behavior where the selected left
nav element is highlighted. This will be re-introduced in subsequent
commits.