feat: experiment - please ignore this commit
This commit introduces the `primaryColors` and `secondaryColors`
state hook at the application root, `App.jsx`. By keeping track of
colors at the root level, we can supply this colors to the `EditPage`
components directly, so we only need to retrieve settings data for
each repository once.

The state hook is stored as an object, where the key is the repo name
and the value is the hex color string. The reason for this is so that
we can keep track of the color of each repo. If we used the state hook
to store only the current repo's color, if the user redirected to another
repo, then that repo's `EditPage` would be rendered with another repo's
colors, which would just be wrong.

We also modify the `EditPage` component to retrieve color settings
data ONLY if the state hook does not have color data for that repo.
No data is fetched if the state hook already has color data for that
repo. Therefore, for each repo, we only need to fetch color data once.
If the user opens up another page on `EditPage`, we would already have
color data for that repo and so we can send it as a prop.