Fix/revamp homepage preview (#172)
* Fix: Changed hero highlights to reflect only title

* Feat: change add new class for homepage add section

* Feat: add watermark to placeholder media

* Feat: add ability to modify site information

* Feat: changed grab cursor to move cursor on draggable components

* Fix: remove site subtitle, description, image fields

We removed site subtitle because it wasn't doing anything, and site image because site-wide changes should be done in settings. Site description has also been temporarily removed since it only shows up as metadata. The label for this section has been changed accordingly.

* Fix: change pointer cursor to default cursor on preview

* Feat: provide default values for homepage sections

* Fix: Fixed bug where homepage could not be saved when changing between dropdown and highlight

Previously, when changing format of the homepage between dropdown and highlights, we had an issue with saving the yaml file, as several fields were left undefined. This has been corrected by setting those fields to the empty string instead.

* Fix: fixed bug where index.md file was not being updated properly

Previously, the index.md file had previously existing fields left as is with an empty string, causing unintended behaviour where switching from dropdown to highlights would cause the dropdown to remain, due to it having higher priority.

The issue was fixed by removing all fields with empty strings when writing to the backend, such that the index.md file could be properly updated.

* Feat: added ability to modify and preview notification

* Feat: added image version of form field

This commit changes the input fields for images to use the image selection modal. This affects the Home page banner, the infopic section on the homepage, and the selection of favicon on the settings page.

* Fix: infopic updates generated site correctly

Previously, changes to the infopic image and alt text were not reflected on the website, as the fields being saved were not the same as the ones being used. The fields have been updated to fix this bug, and additionally a validator has been added to the infopic section.

* Fix: Text overflow in homepage preview

This issue was caused by the fixed height given for the body, causing the child elements to overflow.

* Fix: remove background from notification close preview

* Feat: added placeholder image if no image selected

* Nit: add spacing between colons and titles

* Feat: changed add new section to use button

* Fix: change default value for url

* Fix: right align buttons and remove placeholder for image fields

* Fix: changed new section creator to have button inline