feat: implement ProtectedRoute HOC for login page
This commit implements the ProtectedRoute HOC for the login page.
This includes setting the login state to true once the user is redirected
to the GitHub OAuth login.

With this change, any attempts made to access the `/` route will be
redirected to the `/sites` route if the user is logged in.