feat: initialize login state to be false
Previously, we attempted to initialize login state using an async
function. However, this resulted in some problems since the state
was not being held - somehow, `isLoggedIn` was not displayed as a
boolean but instead as a function in the React browser devtools.

This commit attempts to fix that by initializing login state to be
false, but to immediately check login state using a useLayoutEffect.