fix: react warning upon logout
Previously, upon logging out, we would get this warning message in
the console:

```
Warning: Cannot update a component from inside the function body of a different component.
```

This commit changes the ProtectedRoute so that `setIsLoggedIn` is
not called within the render prop of the Route component inside
ProtectedRoute, but is instead set to false upon loading of the
Home page component.