feat: clear cookie when we set login state to false
When we get a 401 error response and automatically log people out,
we need to clear their cookies as well so that react router doesn't
confused (this is because we initialize our login state based on
whether the cookie exists. therefore it's important to clear the
cookie when we actually logout). This commit makes the axios call to
clear the `isomercms` cookie when we set login state to be false.