fix: define axios interceptor outside of useEffect
Currently, we put the cookie validation useLayoutEffect function
in a try-catch block because the axios interceptor doesn't catch
the 401 returned by the `validate-cookie` endpoint. This is because
the axios instance is used to hit the endpoint BEFORE the interceptor
is defined. In any case, there is no need to define the interceptor
within a useEffect block.

This commit removes the definition of the interceptor from the useEffect
block and puts it at the top of the app.