feat: create HOC route wrapper with redirect rules
This commit creates a HOC which wraps around React Router routes
to help enforce redirect rules. The rules are as follows:

Case 1: User is logged in
- If the user tries to access the `/` page, then they should be
redirected to the login page ONLY if they access it by clicking the
sign out button. Otherwise, we redirect them to the `/sites` page
since we want the user to logout only if they click on the logout button
- If the user tries to access any other page, we allow them to go to
that page

Case 2: User is logged out
- If the user tries to access the `/` page, we will simply render
the login component
- If the user tries to access any other page, we will redirect them
to the login component