feat: initialize login state by checking whether cookie exists
Previously, we initialized login state to be false. The downside of this
is that when you enter a random isomerCMS url into the url bar, due to
the routing rules in place, the user might be shown a flicker of the
login page first before being redirected to the correct page.

However, if we detect that the cookie exists, we can initialize login state
to be true. If upon cookie validation, the login state is shown to be
false, the app can then redirect back to the login page. In the future,
we will also show a modal which indicates that we're automatically logging
the user out.

This is made possible by the `cookieChecker` util function introduced
in this commit. Previously, we did not think it was possible to check for
whether the isomercms cookie exists, since we set it to be httponly, which
prevents javascript from reading the cookie or even detecting whether it
exists. However, there is a trick which allows us to detect whether the
cookie exists, by attempting to write to the same namespace. If the cookie
does not exist, we will be allowed to create a cookie with that name, but
if it does, then we will be blocked from doing so.