feat: use axios interceptor to log user out upon 401 response
This commit adds an axios interceptor which logs the user out when
a 401 response is received from the backend. Since the 401 response
is only received when the JWT is not present, or if it has expired,
this interceptor enables us to create an auto-logout functionality.

The interceptor is defined WITHIN the App component so that it can
set login state to be falsed when an error response is received.