bug: set logout state when clicking logout button
Currently, when we click the logout button when we're logged in,
the isLoggedIn hook's value remains at `true` because we have not
set it. This commit fixes that by using React Context to pass the
setLogoutState function to the Header component directly, instead
of drilling props all the way down. This is the recommended way
of passing a callback down to a descendant component - refer to React's
official docs: https://reactjs.org/docs/hooks-faq.html#how-to-avoid-passing-callbacks-down