chore: reset all state hooks if logged out
Currently, if a user is automatically logged out, the shouldBlockNavigation
and shouldLogout hooks stay stuck at `true`. Since the useEffects are
dependent on changes in state, and the axios interceptors always
set shouldBlockNavigation to `true`, if the current state is true, and
the interceptor tries to set it to true, there is no state change
and the useEffects will not be triggered. This commit helps to reset
hook states to the original `false` values, if logged out.

The commit also ensures that the setLogoutState function, which is
triggered when the user clicks the logout button, is updated so that
all state is reset together as well.