fix: set login state to be true AFTER receiving cookie
The bug in question is that the login state was set to `true` at the START of the
OAuth process, before we received the callback from GitHub, and redirected the frontend.
This caused a redirection to '/sites' by our own routing logic, not by the callback flow.
Since this occurred before we received the access token in the callback flow, naturally
we get a 401 error since we don't have the access token necessary to retrieve GitHub
resources, and this results in an auto-logout.

This commit fixes that by setting login state to be true only AFTER we receive the
access token from the cookie. This is facilitated by:
1. Changing the frontend redirect URL to `/auth`
2. Creating a new layout, AuthCallback, which sets login state to be true whenever it
is mounted, and then redirect to `/sites`

Additionally, we remove the redirection chunk where the path is '/' but the traffic
did NOT come from the log out button. This logic is now taken care of directly in the
routing layer at App.jsx. Basically, we redirect all unspecified URLs as well as the '/'
path to '/sites' if logged in, and redirect it to the Home component at '/' otherwise.