feat: (hack) filter '/auth' invocations from the URL bar
Currently, our backend redirects the frontend URL to '/auth', which
is where we set login state to true after receiving the access token.

However, the '/auth' path is also accessible via the URL bar, instead
of just from the backend redirection. When one enters the '/auth' path
into the URL bar, the following things happen:
- login state is set to 'true', redirected to '/sites'
- however, since we haven't actually received a token, we are automatically
logged out upon reaching '/sites'. Note that we do see the '/sites' page,
although it is blank and contains no content.

It is not great UI to let users see this, however, so this commit attempts
to hide this from the user. We do so by using a specific string, 'isomercms-backend',
which the backend passes as a query string. We then filter out any path beginning
with '/auth' that does not have the same auth context string. This means that
if soemone types the following into the URL bar, they will not be displayed
the ugly, empty '/sites' page:
- '/auth'
- '/auth?abc'
- '/auth?isomercms'

but if they type the path '/auth?isomercms-backend', then they'll see
the ugly, empty '/sites' page. While this is hacky, this isn't an actual
security vulnerability so we don't need to be too concerned for now.