refactor: store accordion elements in separate array
Previously, when constructing the accordion, the order in which we
generated the accordion elements was jumbled up since we used a map
to return the individual elements, but some components (third nav elements)
needed to be accumulated in a separate array (the thirdNavElements
array). This led to situations where we needed to first push to the
thirdNavElements array before returning the third nav header element
in the map, even though the third nav header element comes first in the
AST.

This commit refactors the code so that all accordion elements are stored
in a separate accordionElements array, rather than being returned from
a map. This allows us to push to build the accordion in the correct order
instead of a jumbled order as it was before.