fix: redirect to /sites only after state change
Previously, on the AuthCallback component, we set the login state
to be true and immediately returned a redirect to the `/sites` page.
The problem with this is that it's possible that the redirection
occurs before the login state is set.

This commit makes the changes to redirect only AFTER login state has
been set by introducing a `shouldRedirectToSites` hook, which the
redirection is dependent on.