feat: add Workspace layout
According to the figma designs, in implementing the hierarchical
file structure for collections, we will be switching to a Workspace
layout which contains the homepage, a bunch of folders leading to
collections, and any unlinked pages.

This commit creates the Workspace layout, and modifies routing so that
the user is redirected to the /sites/:siteName/workspace route which
leads to the Workspace component. It also implements basic elements of
the Workspace component, such as the Homepage folder card.