fix: FolderCard CSS bug
This commit fixes the FolderCard CSS bug which causes neighboring
FolderCard components to be rendered vertically instead of side-by-side.
The cause of the bug is that the following two divs which FolderCard
is currently wrapped in:

<div className={contentStyles.folderContainerBoxes}>
   <div className={contentStyles.boxesContainer}>
     ...
   </div>
</div>

should not be a part of the FolderCard component, since those divs
act as containers for FolderCard components.

This commit also generalizes the component to deal with resources,
collections, and homepage type pages so that the correct link is
generated for each type of page.

Finally, it also updates the component's proptypes.