fix: missing <hr /> line bug
This commit fixes a bug where the segment divider <hr /> disappears
from view after I added the collections and pages sections to the
Workspace layout. The reason for this can be found in the following
article: https://dev.to/alvaromontoro/the-disappearing-line-a-css-mystery-3e35

"When using FlexBox, the elements inside the flex container behave
differently than how they are originally defined. Some properties
like float, clear, or display do not apply in the same way.

Browsers display <hr> as an empty block element with automatic horizontal
margins (that will center it if a width is specified). But when the <hr>
becomes a flex item, the automatic width is pushed to 0. Disappearing
from the view."

Therefore, to fix this, I wrapped the <hr /> component in a div
and specified the div width to be 100%.