feat: separate unlinked pages into a separate component
This commit creates a new component, CollectionPagesSection, which
is used to house a collection of components. This is to be used
in each collection layout, and in this commit, is used to house
unlinked pages as well. This component will be developed in further
commits to add the functionality it needs.