fix: bug in ComponentSettingsModal changeHandler
In the change handler, we do error validation to ensure that the user
doesn't choose a title which clashes with another page in the same
level (collection or unlinked pages). To do this, we must first
exclude the file's own title from consideration. However, there
is a bug in this exclusion process currently since we compare against
this.state.fileName, which does not exist.

This commit fixes that by making the correct comparison to the
fileName prop.