feat: modify CategoryPages to dynamically set back button in Header
Since the CategoryPages component will be used for both collections
as well as resource categories, we need the back button in the header
to change dynamically depending on the route. This commit creates
a util function which helps to set the correct back button text and
url depending on the path used to access CategoryPages.