fix: bug in ComponentSettingsModal category folder selection
Currently, we use the output of isCategoryDropdownDisabled to
determine whether the category folder dropdown should be disabled
or not. However, we pass the `category` state into the function,
which causes the dropdown to be disabled once it's selected, preventing
users from changing their choice. This should be replaced with the
`prevCategory` state instead.