Feat/collections layout (#206)
* feat: update FolderCard appearance

This commit updates the FolderCard component's appearance. The main
change is to make folder card's flex direction row instead of column,
since all the elements should be in the same row. Other changes are:
- allow for change in icon from homepage icon to folder icon
- use <span> instead of <h1> for text content

* feat: add Workspace layout

According to the figma designs, in implementing the hierarchical
file structure for collections, we will be switching to a Workspace
layout which contains the homepage, a bunch of folders leading to
collections, and any unlinked pages.

This commit creates the Workspace layout, and modifies routing so that
the user is redirected to the /sites/:siteName/workspace route which
leads to the Workspace component. It also implements basic elements of
the Workspace component, such as the Homepage folder card.

* fix: FolderCard CSS bug

This commit fixes the FolderCard CSS bug which causes neighboring
FolderCard components to be rendered vertically instead of side-by-side.
The cause of the bug is that the following two divs which FolderCard
is currently wrapped in:

<div className={contentStyles.folderContainerBoxes}>
   <div className={contentStyles.boxesContainer}>
     ...
   </div>
</div>

should not be a part of the FolderCard component, since those divs
act as containers for FolderCard components.

This commit also generalizes the component to deal with resources,
collections, and homepage type pages so that the correct link is
generated for each type of page.

Finally, it also updates the component's proptypes.

* feat: improve file name prettifier to capitalize first letter

This commit improves the file name prettifier by capitalizing the
first letter of each token.

* chore: misc. minor changes

* feat: update Workspace layout to display pages and collections

This commit adds the OverviewCard components to render pages in the
Workspace layout

* fix: missing <hr /> line bug

This commit fixes a bug where the segment divider <hr /> disappears
from view after I added the collections and pages sections to the
Workspace layout. The reason for this can be found in the following
article: https://dev.to/alvaromontoro/the-disappearing-line-a-css-mystery-3e35

"When using FlexBox, the elements inside the flex container behave
differently than how they are originally defined. Some properties
like float, clear, or display do not apply in the same way.

Browsers display <hr> as an empty block element with automatic horizontal
margins (that will center it if a width is specified). But when the <hr>
becomes a flex item, the automatic width is pushed to 0. Disappearing
from the view."

Therefore, to fix this, I wrapped the <hr /> component in a div
and specified the div width to be 100%.

* feat: separate unlinked pages into a separate component

This commit creates a new component, CollectionPagesSection, which
is used to house a collection of components. This is to be used
in each collection layout, and in this commit, is used to house
unlinked pages as well. This component will be developed in further
commits to add the functionality it needs.

* chore: add proptypes to Workspace layout

* feat: add a NewCollectionPages component

This commit adds a NewCollectionPages component which displays
all the pages in any given collection in the repo. Again, more
functionality will be added in subsequent commits - this is simply
to get the skeleton out.

* feat: update Header back buttons

This commit edits the following layouts so that the `Back` button
in the Header component redirects to the proper layout:
- Collections

* fix: content.componentIcon styles

Previously, I mistakenly copied the .componentInfo class into the
.componentInfo sass block, making it a child of .componentInfo when
in fact it should be a sibling. This commit reverts that change.

* feat: add ComponentSettingsModal to CollectionPagesSection

This commit adds functionality to the settings button - now when you
click on the gear, it will trigger the ComponentSettingsModal where
you can modify information about the page itself.

* fix: revamp title generation for OverviewCard

This commit accomplishes a few things:
- properly display title for collection pages, since previously we
blindly used prettifyPageFileName for all pages, but collection
pages need to use prettifyCollectionPageFileName
- generate title from the `fileName` prop within the OverviewCard
component itself

* feat: update header back button info for Resources layout

* fix: logic for ComponentSettingsModal category dropdown

This commit fixes a hole in the logic of deciding whether we should
disable the category dropdown or not. Currently, we allow users
to select the dropdown when creating a new page inside a category/
collection when we don't actually want them to do so. This commit
fixes that.

* chore: add react key for dropdown option elements

* chore: minor adjustment to modal styles for balance

* fix: bug in ComponentSettingsModal changeHandler

In the change handler, we do error validation to ensure that the user
doesn't choose a title which clashes with another page in the same
level (collection or unlinked pages). To do this, we must first
exclude the file's own title from consideration. However, there
is a bug in this exclusion process currently since we compare against
this.state.fileName, which does not exist.

This commit fixes that by making the correct comparison to the
fileName prop.

* chore: remove old collectionpages component

* chore: rename NewCollectionPages to CategoryPages for clarity

* chore: change back button on header to Sites for Resources layout

* feat: highlight sidebar tabs for CategoryPages layout

Currently, in the CategoryPages layout, the sidebar is there but
none of the tabs are highlighted. Since the CategoryPages layout
is accessed from either the Workspace or Resources layout, we want
to highlight either the `My Workspace` or `Resources` tab in the
sidebar.

* feat: modify CategoryPages to dynamically set back button in Header

Since the CategoryPages component will be used for both collections
as well as resource categories, we need the back button in the header
to change dynamically depending on the route. This commit creates
a util function which helps to set the correct back button text and
url depending on the path used to access CategoryPages.

* fix: bug in ComponentSettingsModal category folder selection

Currently, we use the output of isCategoryDropdownDisabled to
determine whether the category folder dropdown should be disabled
or not. However, we pass the `category` state into the function,
which causes the dropdown to be disabled once it's selected, preventing
users from changing their choice. This should be replaced with the
`prevCategory` state instead.

* fix: cleanup conditional rendering logic for CollectionPagesSection

Previously, the CollectionPagesSection component would render the
`Add New Page` button 100% of the time, and only render other pages
if they existed. This commit fixes that by rendering the `Add New
Page` button together with the other pages when we receive them from
the API call response.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>