Fix: Prevent resource page from showing loading resources
This commit fixes a bug where the main resources page would always show loading resources. This was due to the check for a positive number of pages in the CollectionPagesSection, however, the Resources page never had any pages, due to the requirement for all resources to have a category. To fix this, we initialize pages in each component to an undefined variable instead of an empty array, to distinguish between loading pages and not having pages.

The logic for showing the Resource Room portion has also been modified to more accurately show loading messages where appropriate.