Feat: add react select library
This commit serves to add the ability for users to create new categories of their own when selecting a category. The default values have been changed for new files to be consistent across different items. In addition, validation has been added to the category field to account for this new change.