Fix: prevent category select from being temporarily active when initialized
This commit fixes a bug in which the category select option would be available for users for a short time when initialized, potentially allowing a user to clicks on it and type something before it gets disabled.

To prevent this, we instead shift the disabled evaluation to CollectionPagesSection instead, so as to immediately set the field to be disabled instead of waiting for the state to be evaluated.

In addition, this commit also fixes a bug in which an unlinked page had the incorrect label applied to its selection