Feat/resource file hierarchy (#207)
* Feat: extend existing CollectionPagesSection and CategoryPages for resources

This commit extends the existing classes in preparation for their use in the revamped resources section, as our resources section is intended to have a similar layout to the existing workspace page.

* Feat: add new page for resource category

* Feat: revamp resource page

This commit revamps the existing resource page to follow a similar layout to the existing one used by workspace.

* Fix: Prevent resource page from showing loading resources

This commit fixes a bug where the main resources page would always show loading resources. This was due to the check for a positive number of pages in the CollectionPagesSection, however, the Resources page never had any pages, due to the requirement for all resources to have a category. To fix this, we initialize pages in each component to an undefined variable instead of an empty array, to distinguish between loading pages and not having pages.

The logic for showing the Resource Room portion has also been modified to more accurately show loading messages where appropriate.

* Rename: Collections to CategoryPages

* Fix: add missing isResource prop to CollectionPagesSection in Resources

* Fix: add isRequired to siteName

* Fix: reload page after creating resource room

* Fix: check for existence of resourceCategories

* Fix: remove isRequired requirement from errorMessage in FormFieldMedia

* Feat: add react select library

This commit serves to add the ability for users to create new categories of their own when selecting a category. The default values have been changed for new files to be consistent across different items. In addition, validation has been added to the category field to account for this new change.

* Fix: update validation for category to allow _ and -

* Fix: reset errors for permalink when toggled

* Nit: remove log statement

* Nit: rename handler for dropdown

* Nit: rename prevCategory to originalCategory

* Fix: prevent category select from being temporarily active when initialized

This commit fixes a bug in which the category select option would be available for users for a short time when initialized, potentially allowing a user to clicks on it and type something before it gets disabled.

To prevent this, we instead shift the disabled evaluation to CollectionPagesSection instead, so as to immediately set the field to be disabled instead of waiting for the state to be evaluated.

In addition, this commit also fixes a bug in which an unlinked page had the incorrect label applied to its selection