feat: refactor ComponentSettingsModal with hooks
This commit tests a refactor of the ComponentSettingsModal into hooks.
There were a few reasons why a refactor was required (not necessarily
into hooks):
1. There was confusion over the `category` variable, which sometimes
referred to the state variable, and sometimes referred to the prop.
To make things more confusing, the prop was also stored as a state
variable called `originalCategory`.
2. `type` and `originalCategory` props were set as state variables
even though they were never modified. They should remain as props
and remain accessible as props.
3. The team is more familiar with hooks, which also simplifies the
definition of functions (no need to rely on `this`), and also
on the usage of props and state (no need to always destructure from
props or `this.state`)

This commit also renames the `date` state variable to `resourceDate`
to be more specific, since normal pages do not have a `date` attribute.