feat: replace third nav title field with a dropdown
This commit replaces the third nav title form field in the
ComponentSettingsModal with the same dropdown component, CreatableSelect,
that was used for the category dropdown. This commit adds functionality
(meaning, you can actually change the third nav of different
files using the dropdown).

However, more work needs to be done for the component to display
the correct default values. This will be accomplished in subsequent commits.