fix: use async version of CreatableSelect for third nav dropdown
Previously, we used the synchronous version of CreatableSelect to
render the dropdown options for the third nav titles. The problem with
this is that the api call to retrieve the third nav titles takes quite
a bit of time. Even though I set the dropdown options as a state hook,
the dropdown did not re-render the component with the full third nav
title options after they were retrieved. I had to close and reopen the
ComponentSettingsModal for them to be displayed.

Therefore, I chose to switch to use AsyncCreatableSelect, which allows
you to pass in a promise which resolves to the dropdown options. This
is perfect since we can just pass in the async function containing the
api call to retrieve third nav titles.

It's important to note that state is stored on the parent component,
CollectionPagesSection, instead of the ComponentSettingsModal itself.
This is so that the third nav title dropdown options can be reused
and passed as props when other pages in the collection need it, instead
of making the same api call again to retrieve the options.

We might consider using this for the category dropdown as well, since
that also involves an async api call on component mount. However, we
have not faced much of a problem with that so far since the API call
for that function returns much faster, so we'll leave it for now.

This commit also modifies both the existing CreatableSelect and the
new AsyncCreatableSelect to use the `value` prop instead of the
`defaultValue` prop. This is because we run into a similar problem as
above if we use `defaultValue`: if the value passed in is a state hook
which is instantiated as an empty string, the component does not update,
meaning that the value rendered on the component stays as an empty string
and does not update to reflect the newly set state hook value.