fix: miscellaneous fixes to form changeHandler
This commit makes two fixes:

1. Create a separate dropdown handler for form dropdown elements.
This is because dropdowns don't need to go through the same validation
process that the other fields need.

2. When constructing the new file name in the validation process
for files of type `page`, it currently assumes that all pages are
simple pages. This commit modifies the process so that a collection
page file name is generated instead, if the file belongs to a collection.