feat: implement file-naming logic
This commit implements the file-naming logic for several situations:
1. When we create a new page in a collection
2. When we move a page that is within a collection in OR out of a
third nav section

This is important because the third nav does not render properly if
the filenames of all pages in the same third nav section are not adjacent to each other alphanumerically. For example, if the file names in a collection are:

1-abc.md
2-def.md
3-ghi.md

If 1-abc.md and 3-ghi.md both have third_nav_title = ‘test’, two separate third nav sections named `test` will show up.

The test cases to check for are as follows:
1. Create new collection page from workspace
2. Create new collection page with third nav from within a collection
3. Create new collection page without third nav from within a collection
4. Change a third nav page to a regular page from within a collection
5. Change a regular page to a third nav page from within a collection
6. Change a third nav page from one third nav section to another (within a collection)
7. Create a new third nav section and assign a regular page in the collection to it

Some of the implementation details:
- To check whether a page’s third nav title value has changed, we introduce a new `originalThirdNav` state hook
- We add the page being created/modified to the bottom of its new group (either a third nav section, or the entire collection itself).

For example, we have the following files in a collection:

1a-abc.md
1b-def.md
2-ghi.md
3-jkl.md

Where 1a-abc.md and 1b-def.md belong to the same third nav ‘test’. If I want to assign a page to the `test` third nav, the file name will be 1c-<title>.md. The group identifier, the 1a, 1b, etc. prefix, will be incremented alphabetically. If the letter `z` is reached, then we append a new letter `a` to the identifier. For example, the identifier would go from 1z to 1za.

Likewise, if I just want to add a new page to the collection, and not in a third nav, then the file name will be 4-<title>.md.

If we are creating a new third nav, we increment the number and set the alphabet to `a`. So for example, the last file could either contain an alphabet identifier or not:
- 3-<title>.md
- 3c-<title>.md

But in both cases, the new third nav file should be 4a-<title>.md